requirements.txt
openai>=1.0.0
langchain>=0.2.0
streamlit==1.31.1
faiss-cpu==1.7.4
mlflow==2.11.1
python-dotenv==1.0.1
fpdf==1.7.2
matplotlib==3.8.3
PyPDF2==3.0.1
pdfplumber==0.10.3
langchain-community>=0.0.24
langchain-openai>=0.1.6
pypdf==5.4.0
pytest
supabase>=0.5.6